requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
sqlite3
asyncio
requests
google-api-python-client
unidecode
yt-dlp
openai